added events exchange

+++ books/src/modules/books/books.controller.ts
@@ -1,16 +1,31 @@
-import { Controller } from '@nestjs/common';
-import { GrpcMethod } from '@nestjs/microservices';
+import { Controller, NotFoundException, Inject } from '@nestjs/common';
+import { GrpcMethod, ClientProxy } from '@nestjs/microservices';
 
 import { Book } from './books.type';
 import { BooksService } from './books.service';
+import { BOOK_CREATED, BOOK_DELETED } from './books.const';
+import { AuthorsService } from '../authors/authors.service';
+import { AUTHORS_CLIENT } from '../authors/authors.const';
 
 @Controller()
 export class BooksController {
-  constructor(private readonly booksService: BooksService) {}
+  constructor(
+    private readonly booksService: BooksService,
+    private readonly authorsService: AuthorsService,
+    @Inject(AUTHORS_CLIENT) private readonly client: ClientProxy
+  ) {}
 
   @GrpcMethod()
-  findAll(): {books: Book[]} {
-    return {books: this.booksService.getBooks()};
+  findAll(): {books: Book[] & {authorFirstName: string, authorLastName: string}[]} {
+    const books = this.booksService.getBooks()
+    return {books: books.map(book => {
+        const author = this.authorsService.findById(book.authorId);
+        return {
+          ...book,
+          authorFirstName: author.firstName,
+          authorLastName: author.lastName
+        }
+      })};
   }
 
   @GrpcMethod()
@@ -20,12 +35,21 @@ export class BooksController {
 
   @GrpcMethod()
   create(createBookDTO: Partial<Book>): Book {
-    return this.booksService.create(createBookDTO);
+    this.validateAuthor(createBookDTO.authorId);
+    const book = this.booksService.create(createBookDTO);
+
+    this.client.emit(BOOK_CREATED, {authorId: book.authorId, bookId: book.id});
+
+    return book;
   }
 
   @GrpcMethod()
   update(data: Partial<Book>): Book {
     const { id, ...updateBookDTO } = data;
+
+    if (updateBookDTO.authorId) {
+      this.validateAuthor(updateBookDTO.authorId);
+    }
 
     return this.booksService.update(+id, updateBookDTO);
   }
@@ -33,7 +57,19 @@ export class BooksController {
   @GrpcMethod()
   delete(data: {id: string}): void {
     const { id } = data;
+    const book = this.booksService.findById(+data.id);
+    if (!book) {
+      throw new NotFoundException(`Book ${id} not found`);
+    }
 
-    return this.booksService.delete(+id);
+    this.booksService.delete(+id);
+    this.client.emit(BOOK_CREATED, {authorId: book.authorId, bookId: book.id});
+  }
+
+  private validateAuthor(authorId: number): void {
+    const author = this.authorsService.findById(authorId);
+    if (!author) {
+      throw new NotFoundException(`Author ${authorId} not found`);
+    }
   }
 }
